+++ politeiawww/cmsdatabase/database.go
@@ -51,8 +51,8 @@ type Invoice struct {
 	Token              string
 	UserID             string
 	Username           string // Only populated when reading from the database
-	Month              uint16
-	Year               uint16
+	Month              uint
+	Year               uint
 	Timestamp          int64
 	Status             cms.InvoiceStatusT
 	StatusChangeReason string
@@ -64,7 +64,7 @@ type Invoice struct {
 	ContractorName     string
 	ContractorLocation string
 	ContractorContact  string
-	ContractorRate     uint16
+	ContractorRate     uint
 	PaymentAddress     string
 
 	LineItems []LineItem      // All line items parsed from the raw invoice provided.
@@ -74,14 +74,14 @@ type Invoice struct {
 // LineItem contains information about the individual line items contained in an
 // invoice coming into or out of the cmsdatabase.
 type LineItem struct {
-	LineNumber   uint16
+	LineNumber   uint
 	InvoiceToken string
 	Type         cms.LineItemTypeT
 	Subtype      string
 	Description  string
 	ProposalURL  string
-	Labor        uint16
-	Expenses     uint16
+	Labor        uint
+	Expenses     uint
 }
 
 // InvoiceChange contains entries for any status update that occurs to a given
